Redesign MatrixReaders to make them extendable
redesign of class AbstractPrimitiveMatrixReader:
moved a lot of code into the base class
delegating access to entries to the according Octave types
OctaveDouble, OctaveBoolean, OctaveInt,...
so that these classes are now very thin and extendable

in AbstractPrimitiveMatrixReader had to add another type parameter D and
T has to extend AbstractGenericMatrix also
which has to be included in the redesign.

Design improvement: could eliminate
AbstractPrimitiveMatrixReader.product which was code copy

Had to turn AbstractGenericMatrix into an OctaveObject
and to add method setPlain(String, int)
also to subclasses.
To these also added setPlain(double, int), setPlain(boolean, int) and so
on.
This is not needed but pretty.

To unify parsing of values as strings,
moved parsing methods to class StringUtil.
This affects SparseBooleanReader, ScalarReader, ComplexScalarReader,
ComplexMatrixReader, BooeanSingleReader.